+++ frontend/src/pages/org/workflow-detail.ts
@@ -11,7 +11,7 @@ import { until } from "lit/directives/until.js";
 import { when } from "lit/directives/when.js";
 import queryString from "query-string";
 
-import type { Crawl, CrawlLog, Seed, Workflow, WorkflowParams } from "./types";
+import type { Crawl, CrawlLog, Seed, Workflow } from "./types";
 
 import { BtrixElement } from "@/classes/BtrixElement";
 import type { Alert } from "@/components/ui/alert";
@@ -29,12 +29,9 @@ import { WorkflowTab } from "@/routes";
 import { deleteConfirmation, noData, notApplicable } from "@/strings/ui";
 import type { APIPaginatedList, APIPaginationQuery } from "@/types/api";
 import { type CrawlState } from "@/types/crawlState";
-import {
-  NewWorkflowOnlyScopeType,
-  type DuplicateWorkflowSettings,
-  type StorageSeedFile,
-} from "@/types/workflow";
+import { type StorageSeedFile } from "@/types/workflow";
 import { isApiError } from "@/utils/api";
+import { settingsForDuplicate } from "@/utils/crawl-workflows/settingsForDuplicate";
 import {
   DEFAULT_MAX_SCALE,
   inactiveCrawlStates,
@@ -2156,7 +2153,7 @@ export class WorkflowDetail extends BtrixElement {
       <btrix-config-details
         .crawlConfig=${this.workflow}
         .seeds=${this.seeds?.items}
-        .seedFile=${this.seedFileTask.value}
+        .seedFile=${this.seedFileTask.value || undefined}
         anchorLinks
       ></btrix-config-details>
     </section>`;
@@ -2312,31 +2309,22 @@ export class WorkflowDetail extends BtrixElement {
   private async duplicateConfig() {
     if (!this.workflow) await this.workflowTask.taskComplete;
 
-    await Promise.all([
-      this.seedsTask.taskComplete,
-      this.seedFileTask.taskComplete,
-    ]);
+    if (this.workflow?.config.seedFileId) {
+      await this.seedFileTask.taskComplete;
+    } else {
+      await this.seedsTask.taskComplete;
+    }
 
     await this.updateComplete;
     if (!this.workflow) return;
 
-    const workflowParams: WorkflowParams = {
-      ...this.workflow,
-      name: this.workflow.name ? msg(str`${this.workflow.name} Copy`) : "",
-    };
+    const settings = settingsForDuplicate({
+      workflow: this.workflow,
+      seeds: this.seeds,
+      seedFile: this.seedFileTask.value ?? undefined,
+    });
 
-    const seeds = this.seeds?.items;
-    const seedFile = this.seedFileTask.value ?? undefined;
-
-    this.navigate.to(`${this.navigate.orgBasePath}/workflows/new`, {
-      scopeType:
-        seedFile || (seeds?.length && seeds.length > 1)
-          ? NewWorkflowOnlyScopeType.PageList
-          : workflowParams.config.scopeType,
-      workflow: workflowParams,
-      seeds,
-      seedFile,
-    } satisfies DuplicateWorkflowSettings);
+    this.navigate.to(`${this.navigate.orgBasePath}/workflows/new`, settings);
 
     this.notify.toast({
       message: msg("Copied settings to new workflow."),
