+++ frontend/src/types/workflow.ts
@@ -1,6 +1,6 @@
 import type { StorageFile } from "./storage";
 
-import { ScopeType, type Seed, type WorkflowParams } from "@/types/crawler";
+import { ScopeType } from "@/types/crawler";
 
 export enum NewWorkflowOnlyScopeType {
   PageList = "page-list",
@@ -20,11 +20,4 @@ export type WorkflowTags = {
 export type StorageSeedFile = StorageFile & {
   firstSeed: string;
   seedCount: number;
-};
-
-export type DuplicateWorkflowSettings = {
-  workflow: WorkflowParams;
-  scopeType?: ScopeType | NewWorkflowOnlyScopeType;
-  seeds?: Seed[];
-  seedFile?: StorageSeedFile;
 };
